move wizard store to $lib/wizard/
