+++ src/routes/wizard/connect-wallet/balance/+page.ts
@@ -1,7 +1,7 @@
 import type { ConnectWizardData } from '../+layout';
 import type { OnChainData } from 'trade-executor/schemas/summary';
 import { get } from 'svelte/store';
-import { wizard } from 'wizard/store';
+import { wizard } from '$lib/wizard/store';
 import { config } from '$lib/wallet/client';
 import { getAccount, getBalance } from '@wagmi/core';
 import { getDenominationTokenBalance } from '$lib/eth-defi/helpers';
