+++ src/routes/wizard/connect-wallet/connect/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { ConnectWizardData } from '../+layout';
-	import { wizard } from 'wizard/store';
+	import { wizard } from '$lib/wizard/store';
 	import { wallet } from '$lib/wallet/client';
 	import ConnectWallet from '$lib/wallet/ConnectWallet.svelte';
