+++ src/routes/wizard/connect-wallet/introduction/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { wizard } from 'wizard/store';
+	import { wizard } from '$lib/wizard/store';
 
 	wizard.toggleComplete('introduction');
 </script>
