+++ src/routes/wizard/connect-wallet/success/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { wizard } from 'wizard/store';
+	import { wizard } from '$lib/wizard/store';
 </script>
 
 <div class="connect-wallet-success">
