+++ src/routes/wizard/deposit/balance/+page.ts
@@ -1,6 +1,6 @@
 import type { DepositWizardData } from '../+layout';
 import { get } from 'svelte/store';
-import { wizard } from 'wizard/store';
+import { wizard } from '$lib/wizard/store';
 import { config } from '$lib/wallet/client';
 import { getAccount, getBalance } from '@wagmi/core';
 import { getDenominationTokenBalance } from '$lib/eth-defi/helpers';
