+++ src/routes/wizard/deposit/tos/+page.ts
@@ -1,7 +1,7 @@
 import type { Abi } from 'viem';
 import type { DepositWizardData } from '../+layout';
 import { get } from 'svelte/store';
-import { wizard } from 'wizard/store';
+import { wizard } from '$lib/wizard/store';
 import { config, wallet } from '$lib/wallet/client';
 import { readContracts } from '@wagmi/core';
 import { getTosVersion } from 'trade-executor/helpers/tos';
