+++ src/routes/wizard/redeem/connect/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { RedeemWizardData } from '../+layout';
-	import { wizard } from 'wizard/store';
+	import { wizard } from '$lib/wizard/store';
 	import { wallet } from '$lib/wallet/client';
 	import ConnectWallet from '$lib/wallet/ConnectWallet.svelte';
